abcnre/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abcnre"
dynamic = ["version"]
description = "Approximate Bayesian Computation with Neural Ratio Estimation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "torch>=1.9.0",
    "matplotlib>=3.3.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.60.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "mypy>=0.900",
    "pre-commit>=2.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0", 
    "nbsphinx>=0.8",
    "pandoc>=1.0",
]
examples = [
    "jupyter>=1.0",
    "seaborn>=0.11",
    "plotly>=5.0",
]

[project.urls]
Homepage = "https://github.com/username/abcnre"
Documentation = "https://abcnre.readthedocs.io"
Repository = "https://github.com/username/abcnre"
Issues = "https://github.com/username/abcnre/issues"

[project.scripts]
abcnre = "abcnre.cli.main:main"

[tool.setuptools.dynamic]
version = {attr = "abcnre._version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=abcnre --cov-report=html --cov-report=term-missing"
